started major refactoring of files to join together functions and classes
